Move resolver registry to the CLI too
Mirrors the plugin-registry move: `dead_cst.cli` now imports
`UvResolver` directly from `dead_cst.contrib.uv`, owns the
`_BUILTIN_RESOLVERS: dict[str, type[PathResolver]]` map, and exposes
`_load_resolver` (which still falls back to the
`dead_cst.resolvers` entry-point group).

`dead_cst.resolvers` no longer re-exports `UvResolver` and drops
`BUILTIN_RESOLVERS` / `load_resolver`; tests and docs adjusted.
