+++ CHANGELOG.md
@@ -108,16 +108,20 @@ two versions.
 - **`-e re:<pattern>` magic prefix** is no longer recognized. Use
   `--entrypoint-regex <pattern>` instead. The plain `-e` flag now
   treats its argument as a file path or FQN literal.
-- **Plugin registry moved into the CLI.** The CLI's `--plugin` flag
-  now resolves names against a map declared in
-  :mod:`dead_cst.cli`, populated by direct imports of every builtin;
-  out-of-tree plugins still register under the `dead_cst.plugins`
-  entry-point group. The public re-exports of contrib plugins from
-  :mod:`dead_cst.plugins` are gone — import them from
-  :mod:`dead_cst.contrib` (or the specific contrib submodule).
-  `dead_cst.plugins.load_plugin` and `dead_cst.plugins.BUILTIN_PLUGINS`
-  are removed; :mod:`dead_cst.contrib` no longer uses a lazy
-  ``__getattr__`` indirection.
+- **Plugin and resolver registries moved into the CLI.** The CLI's
+  `--plugin` and `--resolver` flags now resolve names against maps
+  declared in :mod:`dead_cst.cli`, populated by direct imports of
+  every builtin; out-of-tree plugins / resolvers still register under
+  the `dead_cst.plugins` / `dead_cst.resolvers` entry-point groups.
+  The public re-exports of contrib plugins from :mod:`dead_cst.plugins`
+  are gone, as is the `UvResolver` re-export from
+  :mod:`dead_cst.resolvers` — import them from :mod:`dead_cst.contrib`
+  (or the specific contrib submodule). `dead_cst.plugins.load_plugin`,
+  `dead_cst.plugins.BUILTIN_PLUGINS`,
+  `dead_cst.resolvers.load_resolver`, and
+  `dead_cst.resolvers.BUILTIN_RESOLVERS` are removed;
+  :mod:`dead_cst.contrib` no longer uses a lazy ``__getattr__``
+  indirection.
 
 ### Changed (breaking)
 - **Dropped the Python ``SymbolGraph`` adjacency facade.**
